feat(GoNode - modulo 4,): Node - desafio
finalizado todos os requisitos

+++ gonode/modulo4Desafio/app/Controllers/Http/EventController.js
@@ -1,12 +1,17 @@
 'use strict'
 
 const Event = use('App/Models/Event')
+const moment = require('moment')
 
 class EventController {
-  async index ({ request }) {
+  async index ({ request, auth }) {
     const { order } = await request.get()
 
-    const event = await Event.query().orderBy('date', order).with('user').fetch()
+    const event = await Event.query()
+      .where('user_id', auth.user.id)
+      .orderBy('date', order)
+      .with('user')
+      .fetch()
 
     return event
   }
@@ -19,15 +24,27 @@ class EventController {
     return event
   }
 
-  show ({ params, request, response, view }) {
-    const event = Event.findOrFail(params.id)
+  async show ({ params, request, response, auth }) {
+    const event = await Event.findOrFail(params.id)
+
+    if (auth.user.id !== event.user_id) {
+      return response.status(400).send({ error: { message: 'Esse evento pertence a outro usuário' } })
+    }
 
     return event
   }
 
-  async update ({ params, request, response }) {
+  async update ({ params, request, response, auth }) {
     const event = await Event.findOrFail(params.id)
     const data = await request.only(['title', 'date', 'place'])
+
+    if (auth.user.id !== event.user_id) {
+      return response.status(400).send({ error: { message: 'Esse evento pertence a outro usuário' } })
+    }
+
+    if (moment(event.date).isBefore(new Date())) {
+      return response.status(400).send({ error: { message: 'Não é possível editar um evento que já passou' } })
+    }
 
     event.merge(data)
     await event.save()
@@ -35,8 +52,16 @@ class EventController {
     return event
   }
 
-  async destroy ({ params, request, response }) {
+  async destroy ({ params, request, response, auth }) {
     const event = await Event.findOrFail(params.id)
+
+    if (auth.user.id !== event.user_id) {
+      return response.status(400).send({ error: { message: 'Esse evento pertence a outro usuário' } })
+    }
+
+    if (moment(event.date).isBefore(new Date())) {
+      return response.status(400).send({ error: { message: 'Não é possível editar um evento que já passou' } })
+    }
 
     event.delete()
   }
